Test `sed` based on functionality to avoid false negative on OS X.
Previously, the presence of `sed` was detected by calling it with the `--version` argument, but that is not implemented across all versions of `sed`. For example, OS X ships with FreeBSD `sed`, so some developers testing locally would incorrectly be instructed to install `sed`, even though it already was installed.

`sed` commands are much more consistent between versions than arguments are, so testing the results of a command is more reliable.

Fixes #52